requirements.txt
anyio==4.4.0
appdirs==1.4.4
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
h11==0.14.0
htmltools==0.5.2
idna==3.7
linkify-it-py==2.0.3
markdown-it-py==3.0.0
mdit-py-plugins==0.4.1
mdurl==0.1.2
packaging==24.1
prompt-toolkit==3.0.36
python-multipart==0.0.9
questionary==2.0.1
requests==2.32.3
setuptools==69.5.1
shiny==0.10.2
sniffio==1.3.1
starlette==0.37.2
tencentcloud-sdk-python-common==3.0.1173
tencentcloud-sdk-python-tms==3.0.1173
typing_extensions==4.12.2
uc-micro-py==1.0.3
urllib3==2.2.2
uvicorn==0.30.1
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
wheel==0.43.0
